move drawing to separate vc
