+++ Camera/EditClipViewController.swift
@@ -16,9 +16,9 @@ class EditClipViewController: UIViewController, UITextFieldDelegate, UIGestureRe
     @IBOutlet weak var clipView: UIView!
     @IBOutlet weak var overlayView: UIView!
     
-    @IBOutlet weak var drawingView: UIView!
-    @IBOutlet weak var drawingImageView: UIImageView!
-    @IBOutlet weak var temporaryDrawingImageView: UIImageView!
+//    @IBOutlet weak var drawingView: UIView!
+//    @IBOutlet weak var drawingImageView: UIImageView!
+//    @IBOutlet weak var temporaryDrawingImageView: UIImageView!
     
     @IBOutlet weak var textInputTextField: UITextField!
     @IBOutlet var textFieldPanGestureRecognizer: UIPanGestureRecognizer!
@@ -37,11 +37,13 @@ class EditClipViewController: UIViewController, UITextFieldDelegate, UIGestureRe
     
     // Drawing shit
     
-    var lastPoint: CGPoint!
-    var strokeWidth: CGFloat!
-    var strokeColor = UIColor.whiteColor()
-    var strokeOpacity: CGFloat!
-    var didSwipe: Bool!
+    var drawingViewController: DrawingViewController!
+//
+//    var lastPoint: CGPoint!
+//    var strokeWidth: CGFloat!
+//    var strokeColor = UIColor.whiteColor()
+//    var strokeOpacity: CGFloat!
+//    var didSwipe: Bool!
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -58,10 +60,15 @@ class EditClipViewController: UIViewController, UITextFieldDelegate, UIGestureRe
         
         // Set up drawing shit
         
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        
+        drawingViewController = storyboard.instantiateViewControllerWithIdentifier("DrawingViewController") as! DrawingViewController
+        drawingViewController.editClipViewController = self
+        
 //        drawingView.hidden = true
         
-        strokeWidth = 8.0
-        strokeOpacity = 1.0
+//        strokeWidth = 8.0
+//        strokeOpacity = 1.0
     }
     
     override func viewWillAppear(animated: Bool) {
@@ -225,93 +232,104 @@ class EditClipViewController: UIViewController, UITextFieldDelegate, UIGestureRe
     
     // Drawing shit
     
-    @IBAction func selectedColor(sender: AnyObject) {
-        print("selected color: \(sender.tag)")
-        
-        switch sender.tag {
-        case 0:
-            strokeColor = UIColor.redColor()
-        case 1:
-            strokeColor = UIColor.greenColor()
-        case 2:
-            strokeColor = UIColor.blueColor()
-        default:
-            strokeColor = UIColor.whiteColor()
-        }
+    @IBAction func tappedDrawButton(sender: AnyObject) {
+        showDrawingView()
     }
     
-    @IBAction func clearDrawing(sender: AnyObject) {
-        self.drawingImageView.image = nil
+    
+    func showDrawingView() {
+        addChildViewController(drawingViewController)
+        self.view.addSubview(drawingViewController.view)
+        self.drawingViewController.didMoveToParentViewController(self)
     }
     
-    override func touchesBegan(touches: Set<UITouch>, withEvent event: UIEvent?) {
-        didSwipe = false
-        
-        if let touch = touches.first {
-            print("touches began")
-            lastPoint = touch.locationInView(self.view)
-        }
-        
-        super.touchesBegan(touches, withEvent: event)
-    }
+//    @IBAction func selectedColor(sender: AnyObject) {
+//        print("selected color: \(sender.tag)")
+//        
+//        switch sender.tag {
+//        case 0:
+//            strokeColor = UIColor.redColor()
+//        case 1:
+//            strokeColor = UIColor.greenColor()
+//        case 2:
+//            strokeColor = UIColor.blueColor()
+//        default:
+//            strokeColor = UIColor.whiteColor()
+//        }
+//    }
     
-    override func touchesMoved(touches: Set<UITouch>, withEvent event: UIEvent?) {
-        didSwipe = true
-        
-        if let touch = touches.first {
-            print("touches moved")
-            
-            let currentPoint = touch.locationInView(self.view)
-            
-            UIGraphicsBeginImageContext(self.view.frame.size)
-            
-            self.temporaryDrawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
-            
-            CGContextMoveToPoint(UIGraphicsGetCurrentContext(), lastPoint.x, lastPoint.y)
-            CGContextAddLineToPoint(UIGraphicsGetCurrentContext(), currentPoint.x, currentPoint.y)
-            CGContextSetLineCap(UIGraphicsGetCurrentContext(), CGLineCap.Round)
-            CGContextSetLineWidth(UIGraphicsGetCurrentContext(), strokeWidth)
-            CGContextSetStrokeColorWithColor(UIGraphicsGetCurrentContext(), strokeColor.CGColor)
-            CGContextSetBlendMode(UIGraphicsGetCurrentContext(), CGBlendMode.Normal)
-            
-            CGContextStrokePath(UIGraphicsGetCurrentContext())
-            self.temporaryDrawingImageView.image = UIGraphicsGetImageFromCurrentImageContext()
-            self.temporaryDrawingImageView.alpha = strokeOpacity
-            
-            UIGraphicsEndImageContext()
-            
-            lastPoint = currentPoint
-        }
-    }
+//    @IBAction func clearDrawing(sender: AnyObject) {
+//        self.drawingImageView.image = nil
+//    }
     
-    override func touchesEnded(touches: Set<UITouch>, withEvent event: UIEvent?) {
-        if !didSwipe {
-            UIGraphicsBeginImageContext(self.view.frame.size)
-            
-            self.temporaryDrawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
-            
-            CGContextSetLineCap(UIGraphicsGetCurrentContext(), CGLineCap.Round)
-            CGContextSetLineWidth(UIGraphicsGetCurrentContext(), strokeWidth)
-            CGContextSetStrokeColorWithColor(UIGraphicsGetCurrentContext(), strokeColor.CGColor)
-            CGContextMoveToPoint(UIGraphicsGetCurrentContext(), lastPoint.x, lastPoint.y)
-            CGContextAddLineToPoint(UIGraphicsGetCurrentContext(), lastPoint.x, lastPoint.y)
-            
-            CGContextStrokePath(UIGraphicsGetCurrentContext())
-            self.temporaryDrawingImageView.image = UIGraphicsGetImageFromCurrentImageContext()
-            
-            UIGraphicsEndImageContext()
-        }
-        
-        UIGraphicsBeginImageContext(self.drawingImageView.frame.size)
-        self.drawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
-        self.temporaryDrawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
-        self.temporaryDrawingImageView.alpha = strokeOpacity
-        
-        self.drawingImageView.image = UIGraphicsGetImageFromCurrentImageContext()
-        self.temporaryDrawingImageView.image = nil
-        
-        UIGraphicsEndImageContext()
-    }
+//    override func touchesBegan(touches: Set<UITouch>, withEvent event: UIEvent?) {
+//        didSwipe = false
+//        
+//        if let touch = touches.first {
+//            print("touches began")
+//            lastPoint = touch.locationInView(self.view)
+//        }
+//        
+//        super.touchesBegan(touches, withEvent: event)
+//    }
+    
+//    override func touchesMoved(touches: Set<UITouch>, withEvent event: UIEvent?) {
+//        didSwipe = true
+//        
+//        if let touch = touches.first {
+//            print("touches moved")
+//            
+//            let currentPoint = touch.locationInView(self.view)
+//            
+//            UIGraphicsBeginImageContext(self.view.frame.size)
+//            
+//            self.temporaryDrawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
+//            
+//            CGContextMoveToPoint(UIGraphicsGetCurrentContext(), lastPoint.x, lastPoint.y)
+//            CGContextAddLineToPoint(UIGraphicsGetCurrentContext(), currentPoint.x, currentPoint.y)
+//            CGContextSetLineCap(UIGraphicsGetCurrentContext(), CGLineCap.Round)
+//            CGContextSetLineWidth(UIGraphicsGetCurrentContext(), strokeWidth)
+//            CGContextSetStrokeColorWithColor(UIGraphicsGetCurrentContext(), strokeColor.CGColor)
+//            CGContextSetBlendMode(UIGraphicsGetCurrentContext(), CGBlendMode.Normal)
+//            
+//            CGContextStrokePath(UIGraphicsGetCurrentContext())
+//            self.temporaryDrawingImageView.image = UIGraphicsGetImageFromCurrentImageContext()
+//            self.temporaryDrawingImageView.alpha = strokeOpacity
+//            
+//            UIGraphicsEndImageContext()
+//            
+//            lastPoint = currentPoint
+//        }
+//    }
+    
+//    override func touchesEnded(touches: Set<UITouch>, withEvent event: UIEvent?) {
+//        if !didSwipe {
+//            UIGraphicsBeginImageContext(self.view.frame.size)
+//            
+//            self.temporaryDrawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
+//            
+//            CGContextSetLineCap(UIGraphicsGetCurrentContext(), CGLineCap.Round)
+//            CGContextSetLineWidth(UIGraphicsGetCurrentContext(), strokeWidth)
+//            CGContextSetStrokeColorWithColor(UIGraphicsGetCurrentContext(), strokeColor.CGColor)
+//            CGContextMoveToPoint(UIGraphicsGetCurrentContext(), lastPoint.x, lastPoint.y)
+//            CGContextAddLineToPoint(UIGraphicsGetCurrentContext(), lastPoint.x, lastPoint.y)
+//            
+//            CGContextStrokePath(UIGraphicsGetCurrentContext())
+//            self.temporaryDrawingImageView.image = UIGraphicsGetImageFromCurrentImageContext()
+//            
+//            UIGraphicsEndImageContext()
+//        }
+//        
+//        UIGraphicsBeginImageContext(self.drawingImageView.frame.size)
+//        self.drawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
+//        self.temporaryDrawingImageView.image?.drawInRect(CGRectMake(0, 0, self.view.frame.size.width, self.view.frame.size.height))
+//        self.temporaryDrawingImageView.alpha = strokeOpacity
+//        
+//        self.drawingImageView.image = UIGraphicsGetImageFromCurrentImageContext()
+//        self.temporaryDrawingImageView.image = nil
+//        
+//        UIGraphicsEndImageContext()
+//    }
 }
